+++ ultralytics/__init__.py
@@ -1,1 +1,5 @@
 __version__ = "8.0.0.dev0"
+
+from ultralytics.yolo.engine.model import YOLO
+
+__all__ = ["__version__", "YOLO"]  # allow simpler import
