+++ ultralytics/nn/modules.py
@@ -24,8 +24,7 @@ from ultralytics.yolo.utils.ops import Profile, make_divisible, non_max_suppress
 from ultralytics.yolo.utils.plotting import Annotator, colors, save_one_box
 from ultralytics.yolo.utils.tal import dist2bbox, make_anchors
 from ultralytics.yolo.utils.torch_utils import copy_attr, smart_inference_mode
-
-from .autobackend import AutoBackend
+from ultralytics.nn.autobackend import AutoBackend
 
 # from utils.plots import feature_visualization TODO
