+++ ultralytics/tests/functional/test_loaders.py
@@ -1,4 +1,4 @@
-from ultralytics.nn import DetectionModel
+from ultralytics.nn.tasks import DetectionModel
 from ultralytics.yolo.utils.checks import check_yaml
