--- ultralytics/yolo/__init__.py
@@ -1,7 +1,0 @@
-from ultralytics.yolo import v8
-
-from .engine.model import YOLO
-from .engine.trainer import BaseTrainer
-from .engine.validator import BaseValidator
-
-__all__ = ["BaseTrainer", "BaseValidator", "YOLO"]  # allow simpler import
