+++ ultralytics/yolo/engine/model.py
@@ -2,8 +2,7 @@ import torch
 import yaml
 
 from ultralytics import yolo  # noqa required for python usage
-from ultralytics.nn import attempt_load_weights
-from ultralytics.nn.tasks import ClassificationModel, DetectionModel, SegmentationModel
+from ultralytics.nn.tasks import ClassificationModel, DetectionModel, SegmentationModel, attempt_load_weights
 # from ultralytics.yolo.data.utils import check_dataset, check_dataset_yaml
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG
 from ultralytics.yolo.utils import LOGGER
