+++ ultralytics/yolo/engine/predictor.py
@@ -30,7 +30,7 @@ from pathlib import Path
 
 import cv2
 
-from ultralytics.nn import AutoBackend
+from ultralytics.nn.autobackend import AutoBackend
 from ultralytics.yolo.data.dataloaders.stream_loaders import LoadImages, LoadScreenshots, LoadStreams
 from ultralytics.yolo.data.utils import IMG_FORMATS, VID_FORMATS, check_dataset, check_dataset_yaml
 from ultralytics.yolo.utils import LOGGER, ROOT, colorstr, ops
@@ -95,7 +95,7 @@ class BasePredictor:
         device = select_device(self.args.device)
         model = model or self.args.model
         self.args.half &= device.type != 'cpu'  # half precision only supported on CUDA
-        model = AutoBackend(model, device=device, dnn=self.args.dnn, fp16=self.args.half)  # NOTE: not passing data
+        model = AutoBackend(model, device=device, dnn=self.args.dnn, fp16=self.args.half)
         stride, pt = model.stride, model.pt
         imgsz = check_imgsz(self.args.imgsz, s=stride)  # check image size
