+++ ultralytics/yolo/engine/validator.py
@@ -4,7 +4,7 @@ import torch
 from omegaconf import OmegaConf
 from tqdm import tqdm
 
-from ultralytics.nn import AutoBackend
+from ultralytics.nn.autobackend import AutoBackend
 from ultralytics.yolo.data.utils import check_dataset, check_dataset_yaml
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG
 from ultralytics.yolo.utils import LOGGER, TQDM_BAR_FORMAT
