+++ ultralytics/yolo/v8/classify/train.py
@@ -1,8 +1,7 @@
 import hydra
 import torch
 
-from ultralytics.nn import get_model
-from ultralytics.nn.tasks import ClassificationModel
+from ultralytics.nn.tasks import ClassificationModel, get_model
 from ultralytics.yolo import v8
 from ultralytics.yolo.data import build_classification_dataloader
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG, BaseTrainer
